Komipuluje dla 3 kategorii ale zle liczy

+++ include/obrabiarka.h
@@ -6,5 +6,5 @@ class WpisyBranzowe;
 
 class Obrabiarka {
 public:
- vector<int> obrob(vector<WpisyBranzowe> wpisy);
+ vector<float> obrob(vector<WpisyBranzowe> wpisy);
 };
